+++ tool_tests/test_slamdone_v751_filter_contract.py
@@ -29,7 +29,7 @@ class SlamDoneV751FilterContractTest(unittest.TestCase):
 
     def test_version_is_v751_or_later(self):
         pubspec = self.read('pubspec.yaml')
-        self.assertRegex(pubspec, r'version: 7\.(?:[6-9]|5\.[1-9])[0-9.]*\+[0-9]+')
+        self.assertRegex(pubspec, r'version: 7\.(?:10|[6-9]|5\.[1-9])[0-9.]*\+[0-9]+')
 
 
 if __name__ == '__main__':